Running external commands in Vcs is now encapsulated by a nested class.
Vcs.Program basically does what run() and friends did, but wraps it up
in a simple object similar to System.Diagnostics.Process; in fact it is
implemented in terms of Process and ProcessStartInfo!

This is much nicer IMHO than cluttering the Vcs ABC with its own
understanding of the task, when C# allows nested classes in a suitable
manor to do it this way.