Added root property to Vcs and suitable constructors for subclasses
The idea being that root is a classes idea of the repository root path.
I've given each class a default constructor (Init should likely become
static and drop the default ctor) and a constructor taking the root. The
Factory has been modified to return objects ready to rock with those new
contractors.